Fix handling of Unicode BOMs in JSONReader.
Because |char| is signed, the upper bits of the BOM bytes were getting 
interpreted as the sign bit. Cast to |uint8| to preserve them for proper
comparison.

BUG=119975
TEST=Try to install the Gmail extension. Do not get JSON error when unpacking.

Review URL: http://codereview.chromium.org/9860035

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@129373 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/json/json_reader.cc
@@ -163,8 +163,9 @@ Value* JSONReader::JsonToValue(const std::string& json, bool check_root,
   // or <0xEF 0xBB 0xBF>, advance the start position to avoid the
   // JSONReader::BuildValue() function from mis-treating a Unicode BOM as an
   // invalid character and returning NULL.
-  if (json.size() >= 3 && start_pos_[0] == 0xEF &&
-      start_pos_[1] == 0xBB && start_pos_[2] == 0xBF) {
+  if (json.size() >= 3 && static_cast<uint8>(start_pos_[0]) == 0xEF &&
+      static_cast<uint8>(start_pos_[1]) == 0xBB &&
+      static_cast<uint8>(start_pos_[2]) == 0xBF) {
     start_pos_ += 3;
   }
